+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2022-04-17
+### Added
+- Java add-two-numbers solution
+- Java last-stone-weight solution
+
+
 ## [0.0.1] - 2022-04-16
 ### Added
 - Java two-numbers solution
+
+